pyproject.toml
[project]
name = "ai-voice-coding-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.9.0",
    "httpx>=0.27.0",
    "langchain>=0.2.11",
    "langchain-community>=0.2.10",
    "langchain-core>=0.2.23",
    "langgraph>=0.1.12",
    "langchain-groq>=0.1.6",
    "numpy==1.26.4",
    "pyaudio>=0.2.13",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
    "sounddevice>=0.4.6",
    "speechrecognition>=3.14.3",
    "typing-extensions>=4.11.0",
]
